# openapi/system_settings.yaml
openapi: 3.0.3
info:
  title: Infron System Settings API
  version: 1.0.0
  description: System and tenant settings management API for Infron-Core cloud management platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GeneralSettings:
      type: object
      properties:
        name:
          type: string
          example: "Infron Cloud Platform"
          maxLength: 255
        description:
          type: string
          example: "Multi-tenant cloud management platform"
          maxLength: 1000
        contactEmail:
          type: string
          format: email
          example: "admin@muxon.example"
        contactPhone:
          type: string
          pattern: "^[+]?[0-9\\s\\-()]+$"
          example: "+1-555-123-4567"
        defaultTimezone:
          type: string
          default: "UTC"
          example: "UTC"
        defaultLocale:
          type: string
          pattern: "^[a-z]{2}-[A-Z]{2}$"
          default: "en-US"
          example: "en-US"
      additionalProperties: false

    # Security Settings
    SecuritySettings:
      type: object
      properties:
        sessionTimeoutMinutes:
          type: integer
          minimum: 5
          maximum: 1440
          default: 60
          example: 60
        maxLoginAttempts:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          example: 5
        lockoutDurationMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          default: 30
          example: 30
        passwordMinLength:
          type: integer
          minimum: 6
          maximum: 128
          default: 8
          example: 8
        passwordRequireUppercase:
          type: boolean
          default: true
          example: true
        passwordRequireLowercase:
          type: boolean
          default: true
          example: true
        passwordRequireDigit:
          type: boolean
          default: true
          example: true
        passwordRequireSymbol:
          type: boolean
          default: false
          example: false
        passwordExpiryDays:
          type: integer
          minimum: 0
          maximum: 365
          default: 90
          example: 90
        apiRateLimitPerMinute:
          type: integer
          minimum: 1
          maximum: 10000
          default: 100
          example: 100
      additionalProperties: false

    # Notification Settings
    NotificationSettings:
      type: object
      properties:
        smtpEnabled:
          type: boolean
          default: false
          example: false
        smtpHost:
          type: string
          example: "smtp.gmail.com"
        smtpPort:
          type: integer
          minimum: 1
          maximum: 65535
          default: 587
          example: 587
        smtpUsername:
          type: string
          example: "noreply@muxon.example"
        smtpPassword:
          type: string
          writeOnly: true
          example: "secret-password"
        smtpFromEmail:
          type: string
          format: email
          example: "noreply@muxon.example"
        smtpUseTls:
          type: boolean
          default: true
          example: true
        slackWebhookUrl:
          type: string
          format: uri
          example: "https://hooks.slack.com/services/..."
        webhookEnabled:
          type: boolean
          default: false
          example: false
      additionalProperties: false

    # IdP Settings (Core - OIDC Only)
    IdpSettings:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
          example: false
        type:
          type: string
          enum: [oidc]
          default: oidc
          example: oidc
        name:
          type: string
          example: "Keycloak"
        issuerUrl:
          type: string
          format: uri
          example: "https://keycloak.example.com/realms/muxon"
        clientId:
          type: string
          example: "muxon-client"
        clientSecret:
          type: string
          writeOnly: true
          example: "client-secret"
        scopes:
          type: string
          default: "openid,profile,email"
          example: "openid,profile,email"
        autoProvisionUsers:
          type: boolean
          default: true
          example: true
        jwksUri:
          type: string
          format: uri
          example: "https://keycloak.example.com/realms/muxon/protocol/openid-connect/jwks"
        userinfoEndpoint:
          type: string
          format: uri
          example: "https://keycloak.example.com/realms/muxon/protocol/openid-connect/userinfo"
      additionalProperties: false

    # Appearance Settings
    AppearanceSettings:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, auto]
          default: light
          example: light
        logoUrl:
          type: string
          format: uri
          example: "https://cdn.example.com/logo.png"
        faviconUrl:
          type: string
          format: uri
          example: "https://cdn.example.com/favicon.ico"
        primaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          default: "#3b82f6"
          example: "#3b82f6"
        secondaryColor:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          default: "#64748b"
          example: "#64748b"
        customCssUrl:
          type: string
          format: uri
          example: "https://cdn.example.com/custom.css"
      additionalProperties: false

    # System Settings (Aggregate)
    SystemSettings:
      type: object
      properties:
        general:
          $ref: '#/components/schemas/GeneralSettings'
        security:
          $ref: '#/components/schemas/SecuritySettings'
        notifications:
          $ref: '#/components/schemas/NotificationSettings'
        idp:
          $ref: '#/components/schemas/IdpSettings'
        appearance:
          $ref: '#/components/schemas/AppearanceSettings'
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-10-01T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-10-01T12:34:56Z"
        updatedBy:
          type: string
          format: uuid
          readOnly: true
          example: "3d2a4598-f255-5137-f0ca-0c8abf159850"
      additionalProperties: false

    # Test Email Request
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: INVALID_IDP_TYPE
        message:
          type: string
          example: Only OIDC is supported in Infron-Core. SAML and other protocols are Nexus (Enterprise) features.
        details:
          type: object
          additionalProperties: true
paths:
...
  /api/v1/tenants/{tenantId}/settings:
    get:
      tags:
        - Tenant Settings
      summary: Get all tenant settings (merged with system defaults)
      operationId: getTenantSettings
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          description: UUID of tenant
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant settings retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '401':
          $ref: './commons.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './commons.yaml#/components/responses/Forbidden'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/Error'
              examples:
                tenantNotFound:
                  value:
                    code: NOT_FOUND
                    message: Tenant not found
    put:
      tags:
        - Tenant Settings
      summary: Update all tenant settings
      operationId: updateTenantSettings
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          description: UUID of tenant
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettings'
      responses:
        '200':
          description: Tenant settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '400':
          $ref: './commons.yaml#/components/responses/BadRequest'
        '401':
          $ref: './commons.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './commons.yaml#/components/responses/Forbidden'
...